+++ src/test/scala/ttt/RunnerSuite.scala
@@ -5,17 +5,17 @@ import java.io.{ByteArrayInputStream, ByteArrayOutputStream}
 
 class RunnerSuite extends FunSuite with Matchers {
   val board = Board.newBoard(9)
-  val firstHuman = new User(Board.firstPlayer)
-  val secondHuman = new User(Board.secondPlayer)
-  val computer = new Computer(Board.secondPlayer)
-  val messenger = new ttt.Messenger.English
+  val firstHuman = new ttt.player.User(Board.firstPlayer)
+  val secondHuman = new ttt.player.User(Board.secondPlayer)
+  val computer = new ttt.player.computer.Computer(Board.secondPlayer)
+  val messenger = new ttt.messenger.English
   val language = "1\n"
   val againstHuman = "1\n"
   val againstComputer = "2\n"
   val threeByTree = "3\n"
   val fourByFour = "4\n"
 
-  def mock(input: String, secondPlayer: Player) = {
+  def mock(input: String, secondPlayer: ttt.player.Player) = {
     lazy val stream = new ByteArrayOutputStream()
     lazy val in = new ByteArrayInputStream(input.getBytes())
